requirements.txt
transformers
torch
Pillow
Flask
torchvision
numpy
tensorflow
keras
